sushi/pyproject.toml
[tool.poetry]
name = "sushi"
version = "0.1.0"
description = ""
authors = ["lamt3 <lamtomoki@gmail.com>"]
readme = "README.md"
packages = [{include = "tuna"}]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.37"
fastapi = "^0.115.8"
asyncpg = "^0.30.0"
uvicorn = "^0.34.0"
requests = "^2.32.3"
facebook-business = "^21.0.5"
greenlet = "^3.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
